Rename RootNavigation.types
