+++ src/DebugScreen/DebugScreen.tsx
@@ -6,7 +6,7 @@
 
 import React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
-import { DebugScreenNavRouteProp } from '../RootNavigation/types';
+import { DebugScreenNavRouteProp } from '../RootNavigation/RootNavigation.types';
 
 export const DebugScreen = ({ route, navigation }: DebugScreenNavRouteProp) => {
   const { testNavParam } = route.params;
